Start implementation of a complete, multiplatform distribution system
Okay, new plan involves both the use of setup.py/pip AND the use of
pyinstaller, because both serve different functions and I think the
combination is a good one.

pip/setuptools serves as the main package handler, but it comes later.
First, there is a folder, exedist, the contains executables compiled
using pyinstaller.  These executables require no installed anything to
run, they will run the complete software.

Next, using setup.py, I create an sdist - this sdist installs the
package as expected, but also includes a consolescript, which puts a
single, small function on the PATH.  This consolescript just executes
__main__.py.run_as_executable() which, currnetly, does nothing, but in
the future will find the correct executable and run it.  Therefore,
using pip install {tar.gz file}, the executable and a small python
script to access it will be installed and accessible on the path.
Voila!  Nothing else needed.  However, pip has the nice feature of going
and installing the package, too, along with any dependencies called for
in install_requires that it can find.  Of course, for some programs,
like those involving Qt, this won't find everything needed.  Someone
will need to go make sure Qt is installed to before python -m {package}
will work.  However, IF Qt is installed, then the full python package is
importable, which is nice!  And if not, at least all the requirements
are handled. And in many cases, like those not requiring Qt, the python
package will be ready to go immediately after running pip install, ALONG
WITH the executable and the path command that runs the executable.

Next up, I need to make sure that the sdist includes the executables in
a way that they can be accessed by __main__.py.run_as_executable(), with
a consistently accessible path so I can use subprocess to call that
path.

Former-commit-id: a564c1a0752085c87c5174e039b49be9d723bad7