Major update to distribution system for executables and Python module
The new distribution system has been selected - the project (and future
projects, likely) will be distributed using a combination of two
methods.

The first will distribute the project as a Python module using
setuptools, wheel, and PyPI.  The module will be available for
installation using pip install either using the wheel file contained in
dist/, or directly from PyPI, where it will be located.  Using the
Python module will require that all packages be installed correctly, and
while pip will install everything it can find in PyPI according to
install_requires, that may not be everything.  The README will contain
any information on what other packages might be required.

The second distribution method will be a set of executables, compiled
using PyInstaller.  The executables will be large and will not require
anything installed in the system - they will be stand-alone files,
contained under executables/{OS}/rnaseqflow_{version}, where {OS} is a
folder of either win, mac_osx, or any other operating system for which
the software is built.

Former-commit-id: f03bf49240fc2223c4e0c64111209659b2876b63