Major changes to structure - further work pending, currently broken
Rather than having a hardcoded workflow, this commit starts version
0.2.0 which defines a workflow as a customizable series of steps to
chain together to process RNAseq data.  The class hierarchy for these
steps is defined.  The structure of the program is also changed to be a
command line utility with an extensive series of arguments.

No stages are actually defined yet, and the previously working workflow
is now broken.  Tests for each stage have not yet been written, but will
be shortly.

Former-commit-id: 0103bd8f263ad8014e9374b3acb431f4645791cd