Allow use of a specific, non-PATH fastq-mcf executable
Testing is hard when your tests are dependent on installed non-Python
executables.  This commit gives a way to bundle the fastq-mcf executable
file instead of expecting it to be on the path by specifying --fastq
argument to be the location of the executable.

Also changes some other arguments and improves path completion

Former-commit-id: 48f68204653f4ea6bb6c2250865f00e73fe08f0b