Increased test coverage, redo docs, removed built docs from git
Small increases in test coverage thanks to actually looking at the test
coverage report.  Some more tests are needed, but it's pretty good for
now.

Documentation has been redrawn and improved, and the build documentation
files are removed from git because Read the Docs builds your
documentation for you from source, which I didn't know before.

Documentation is in a good state although it needs more in the command
line interface and introduction pages.  But that will be work for
another branch.

Former-commit-id: fd69d9477958b8d2a2a71d551fb5db0801fae788